requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
langchain_sqlserver
langchain_community
langchain_openai
azure-search-documents
sqlalchemy
pyodbc
azure-identity
pandas